Added debug output for ExCO mode change to Freeze. Updated Enums and code format.

+++ source/DSES/ExecutionConfiguration.cpp
@@ -647,8 +647,8 @@ bool ExecutionConfiguration::wait_for_update() // RETURN: -- None.
    if ( this->any_remotely_owned_subscribed_init_attribute() ) {
 
       long long    wallclock_time;
-      SleepTimeout print_timer( federate->wait_status_time );
-      SleepTimeout sleep_timer( THLA_LOW_LATENCY_SLEEP_WAIT_IN_MICROS );
+      SleepTimeout print_timer( (double)federate->wait_status_time );
+      SleepTimeout sleep_timer( (long)THLA_LOW_LATENCY_SLEEP_WAIT_IN_MICROS );
 
       // Wait for the data to arrive.
       while ( !this->is_changed() ) {
