+++ source/IMSim/ExecutionControl.cpp
@@ -747,7 +747,7 @@ Simulation has started and is now running...%c",
 FederateJoinEnum ExecutionControl::determine_if_late_joining_or_restoring_federate()
 {
    long long    wallclock_time;
-   SleepTimeout print_timer( federate->wait_status_time );
+   SleepTimeout print_timer( (double)federate->wait_status_time );
    SleepTimeout sleep_timer;
 
    // Block until we have determined if we are a late joining federate.
@@ -1191,7 +1191,7 @@ void ExecutionControl::wait_for_all_multiphase_init_sync_points()
            && ( sp->label.compare( IMSim::SIM_CONFIG_SYNC_POINT ) != 0 ) ) {
 
          long long    wallclock_time;
-         SleepTimeout print_timer( federate->wait_status_time );
+         SleepTimeout print_timer( (double)federate->wait_status_time );
          SleepTimeout sleep_timer;
 
          // Wait for the federation to synchronized on the sync-point.
@@ -2324,7 +2324,7 @@ void ExecutionControl::exit_freeze()
          federate->register_generic_sync_point( IMSim::FEDRUN_SYNC_POINT );      // this tells federates to go to run
 
          long long    wallclock_time;
-         SleepTimeout print_timer( federate->wait_status_time );
+         SleepTimeout print_timer( (double)federate->wait_status_time );
          SleepTimeout sleep_timer;
 
          while ( !this->pause_sync_pts.check_sync_points( this->checktime ) ) {
@@ -2632,7 +2632,7 @@ bool ExecutionControl::is_save_initiated()
       federate->register_generic_sync_point( IMSim::FEDSAVE_SYNC_POINT );
 
       long long    wallclock_time;
-      SleepTimeout print_timer( federate->wait_status_time );
+      SleepTimeout print_timer( (double)federate->wait_status_time );
       SleepTimeout sleep_timer;
 
       while ( !federate->initiate_save_flag ) { // wait for federation to be synced
