+++ source/TrickHLA/SyncPntListBase.cpp
@@ -287,7 +287,7 @@ void SyncPntListBase::wait_for_list_synchronization(
 {
    bool         achieved, valid_and_not_achieved;
    long long    wallclock_time;
-   SleepTimeout print_timer( federate->wait_status_time );
+   SleepTimeout print_timer( (double)federate->wait_status_time );
    SleepTimeout sleep_timer;
 
    // Iterate through this SyncPntList's synchronization point list.
@@ -806,7 +806,7 @@ bool SyncPntListBase::wait_for_sync_point_announcement(
 
       bool         print_summary = DebugHandler::show( DEBUG_LEVEL_4_TRACE, DEBUG_SOURCE_FEDERATE );
       long long    wallclock_time;
-      SleepTimeout print_timer( federate->wait_status_time );
+      SleepTimeout print_timer( (double)federate->wait_status_time );
       SleepTimeout sleep_timer;
 
       // Wait for the sync-point to be announced.
@@ -948,7 +948,7 @@ bool SyncPntListBase::wait_for_synchronization(
       bool         print_summary = DebugHandler::show( DEBUG_LEVEL_2_TRACE, DEBUG_SOURCE_FEDERATE );
       bool         synchronized;
       long long    wallclock_time;
-      SleepTimeout print_timer( federate->wait_status_time );
+      SleepTimeout print_timer( (double)federate->wait_status_time );
       SleepTimeout sleep_timer;
 
       // Wait for the federation to synchronize on the sync-point.
